requirements.txt
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
requests>=2.32.0
python-dateutil>=2.9.0
python-docx>=1.1.2
flask>=3.0.0
